requirements.txt
Django==3.0.5
djangorestframework==3.11.0
django-oauth-toolkit==1.3.2
django-allauth==0.41.0
asgiref==3.2.7
pip==20.0.2
pytz==2019.3
setuptools==46.1.3
sqlparse==0.3.1
django-filter==2.2.0
drf-yasg==1.17.1
Pillow==7.1.1
django-jsonfield==1.4.0
